EY-3796 fjern gamle trygdetid API
