--- apps/etterlatte-trygdetid/src/main/kotlin/trygdetid/NyTrygdetidRoutes.kt
@@ -1,270 +1,0 @@
-package no.nav.etterlatte.trygdetid
-
-import io.ktor.http.HttpStatusCode
-import io.ktor.server.application.ApplicationCall
-import io.ktor.server.application.call
-import io.ktor.server.request.receive
-import io.ktor.server.response.respond
-import io.ktor.server.routing.Route
-import io.ktor.server.routing.delete
-import io.ktor.server.routing.get
-import io.ktor.server.routing.post
-import io.ktor.server.routing.route
-import io.ktor.util.pipeline.PipelineContext
-import no.nav.etterlatte.libs.common.feilhaandtering.GenerellIkkeFunnetException
-import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
-import no.nav.etterlatte.libs.common.trygdetid.DetaljertBeregnetTrygdetidResultat
-import no.nav.etterlatte.libs.common.trygdetid.MigreringOverstyringDto
-import no.nav.etterlatte.libs.common.trygdetid.StatusOppdatertDto
-import no.nav.etterlatte.libs.common.trygdetid.TrygdetidGrunnlagDto
-import no.nav.etterlatte.libs.common.trygdetid.TrygdetidOverstyringDto
-import no.nav.etterlatte.libs.common.trygdetid.TrygdetidYrkesskadeDto
-import no.nav.etterlatte.libs.ktor.brukerTokenInfo
-import no.nav.etterlatte.libs.ktor.route.BEHANDLINGID_CALL_PARAMETER
-import no.nav.etterlatte.libs.ktor.route.behandlingId
-import no.nav.etterlatte.libs.ktor.route.uuid
-import no.nav.etterlatte.libs.ktor.route.withBehandlingId
-import no.nav.etterlatte.libs.ktor.route.withParam
-import no.nav.etterlatte.trygdetid.klienter.BehandlingKlient
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import java.util.UUID
-
-private const val TRYGDETIDID_CALL_PARAMETER = "trygdetidId"
-
-fun PipelineContext<*, ApplicationCall>.trygdetidId(): UUID {
-    return try {
-        this.call.parameters[TRYGDETIDID_CALL_PARAMETER]?.let { UUID.fromString(it) }!!
-    } catch (e: Exception) {
-        throw UgyldigForespoerselException(
-            "MANGLER_TRYGDETID_ID",
-            "Kunne ikke lese ut parameteret trygdetidId",
-        )
-    }
-}
-
-private val logger: Logger = LoggerFactory.getLogger("TrygdetidV2Route")
-
-fun Route.trygdetidV2(
-    trygdetidService: NyTrygdetidService,
-    behandlingKlient: BehandlingKlient,
-) {
-    route("/api/trygdetid_v2/{$BEHANDLINGID_CALL_PARAMETER}") {
-        get {
-            withBehandlingId(behandlingKlient) {
-                logger.info("Henter trygdetid for behandling $behandlingId")
-                val trygdetider = trygdetidService.hentTrygdetiderIBehandling(behandlingId, brukerTokenInfo)
-                if (trygdetider.isNotEmpty()) {
-                    call.respond(trygdetider.map { it.toDto() })
-                } else {
-                    call.respond(HttpStatusCode.NoContent)
-                }
-            }
-        }
-
-        post {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                logger.info("Oppretter trygdetid(er) for behandling $behandlingId")
-                trygdetidService.opprettTrygdetiderForBehandling(behandlingId, brukerTokenInfo)
-                call.respond(
-                    trygdetidService.hentTrygdetiderIBehandling(behandlingId, brukerTokenInfo)
-                        .map { it.toDto() },
-                )
-            }
-        }
-
-        post("oppdater-opplysningsgrunnlag") {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                logger.info("Oppdaterer opplysningsgrunnlag på trygdetider for behandling $behandlingId")
-
-                trygdetidService.oppdaterOpplysningsgrunnlagForTrygdetider(behandlingId, brukerTokenInfo)
-                call.respond(
-                    trygdetidService.hentTrygdetiderIBehandling(behandlingId, brukerTokenInfo)
-                        .minBy { it.ident }
-                        .toDto(),
-                )
-            }
-        }
-
-        post("overstyr") {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                logger.info("Oppdater trygdetid (overstyring) for behandling $behandlingId")
-                val trygdetidOverstyringDto = call.receive<TrygdetidOverstyringDto>()
-
-                trygdetidService.overstyrNorskPoengaaarForTrygdetid(
-                    trygdetidOverstyringDto.id,
-                    behandlingId,
-                    trygdetidOverstyringDto.overstyrtNorskPoengaar,
-                    brukerTokenInfo,
-                )
-                call.respond(
-                    trygdetidService.hentTrygdetidIBehandlingMedId(
-                        behandlingId,
-                        trygdetidOverstyringDto.id,
-                        brukerTokenInfo,
-                    )!!.toDto(),
-                )
-            }
-        }
-
-        post("yrkesskade") {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                logger.info("Oppdater trygdetid (yrkesskade) for behandling $behandlingId")
-                val trygdetidYrkesskadeDto = call.receive<TrygdetidYrkesskadeDto>()
-
-                trygdetidService.setYrkesskade(
-                    trygdetidYrkesskadeDto.id,
-                    behandlingId,
-                    trygdetidYrkesskadeDto.yrkesskade,
-                    brukerTokenInfo,
-                )
-
-                call.respond(
-                    trygdetidService.hentTrygdetidIBehandlingMedId(
-                        behandlingId,
-                        trygdetidYrkesskadeDto.id,
-                        brukerTokenInfo,
-                    )!!.toDto(),
-                )
-            }
-        }
-
-        route("{$TRYGDETIDID_CALL_PARAMETER}") {
-            post("grunnlag") {
-                withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                    logger.info("Legger til trygdetidsgrunnlag for behandling $behandlingId")
-                    val trygdetidgrunnlagDto = call.receive<TrygdetidGrunnlagDto>()
-
-                    trygdetidService.lagreTrygdetidGrunnlagForTrygdetidMedIdIBehandling(
-                        behandlingId,
-                        trygdetidId(),
-                        trygdetidgrunnlagDto.toTrygdetidGrunnlag(brukerTokenInfo),
-                        brukerTokenInfo,
-                    )
-                    call.respond(
-                        trygdetidService.hentTrygdetidIBehandlingMedId(
-                            behandlingId,
-                            trygdetidId(),
-                            brukerTokenInfo,
-                        )!!.toDto(),
-                    )
-                }
-            }
-
-            delete("/grunnlag/{trygdetidGrunnlagId}") {
-                withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                    withParam("trygdetidGrunnlagId") { trygdetidGrunnlagId ->
-                        logger.info("Sletter trygdetidsgrunnlag for behandling $behandlingId")
-                        trygdetidService.slettTrygdetidGrunnlagForTrygdetid(
-                            behandlingId,
-                            trygdetidId(),
-                            trygdetidGrunnlagId,
-                            brukerTokenInfo,
-                        )
-                        call.respond(
-                            trygdetidService.hentTrygdetidIBehandlingMedId(
-                                behandlingId,
-                                trygdetidId(),
-                                brukerTokenInfo,
-                            )!!.toDto(),
-                        )
-                    }
-                }
-            }
-        }
-
-        post("/oppdater-status") {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) { behandlingId ->
-                val statusOppdatert =
-                    trygdetidService.sjekkGyldighetOgOppdaterBehandlingStatus(behandlingId, brukerTokenInfo)
-                call.respond(HttpStatusCode.OK, StatusOppdatertDto(statusOppdatert))
-            }
-        }
-
-        post("/kopier/{forrigeBehandlingId}") {
-            withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                logger.info("Oppretter kopi av forrige trygdetider for behandling $behandlingId")
-                val forrigeBehandlingId = call.uuid("forrigeBehandlingId")
-                trygdetidService.kopierSisteTrygdetidberegninger(behandlingId, forrigeBehandlingId, brukerTokenInfo)
-                call.respond(HttpStatusCode.OK)
-            }
-        }
-
-        route("/migrering") {
-            post {
-                withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                    logger.info("Migrering overstyrer trygdetid for behandling $behandlingId")
-
-                    val overstyringDto = call.receive<MigreringOverstyringDto>()
-
-                    trygdetidService.overstyrBeregnetTrygdetidForAvdoed(
-                        behandlingId,
-                        overstyringDto.ident,
-                        overstyringDto.detaljertBeregnetTrygdetidResultat,
-                    )
-                    call.respond(
-                        trygdetidService.hentTrygdetiderIBehandling(behandlingId, brukerTokenInfo)
-                            .first { trygdetid -> trygdetid.ident == overstyringDto.ident }
-                            .toDto(),
-                    )
-                }
-            }
-
-            post("/manuell/opprett") {
-                withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                    logger.info("Oppretter trygdetid med overstyrt for behandling $behandlingId")
-                    trygdetidService.opprettOverstyrtBeregnetTrygdetid(behandlingId, brukerTokenInfo)
-                    call.respond(HttpStatusCode.OK)
-                }
-            }
-
-            route("/{$TRYGDETIDID_CALL_PARAMETER}") {
-                post("/manuell/lagre") {
-                    withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                        logger.info("Oppdaterer trygdetid med overstyrt for behandling $behandlingId")
-                        val beregnetTrygdetid = call.receive<DetaljertBeregnetTrygdetidResultat>()
-
-                        val eksisterendeTygdetid =
-                            trygdetidService.hentTrygdetidIBehandlingMedId(
-                                behandlingId,
-                                trygdetidId(),
-                                brukerTokenInfo,
-                            ) ?: throw GenerellIkkeFunnetException()
-
-                        val trygdetid =
-                            trygdetidService.overstyrBeregnetTrygdetidForAvdoed(
-                                behandlingId,
-                                eksisterendeTygdetid.ident,
-                                beregnetTrygdetid,
-                            )
-
-                        behandlingKlient.settBehandlingStatusTrygdetidOppdatert(trygdetid.behandlingId, brukerTokenInfo)
-
-                        call.respond(trygdetidService.hentTrygdetidIBehandlingMedId(behandlingId, trygdetidId(), brukerTokenInfo)!!.toDto())
-                    }
-                }
-
-                post("/uten_fremtidig") {
-                    withBehandlingId(behandlingKlient, skrivetilgang = true) {
-                        logger.info("Beregn trygdetid uten fremtidig trygdetid for behandling $behandlingId")
-
-                        val trygdetid = trygdetidService.hentTrygdetidIBehandlingMedId(behandlingId, trygdetidId(), brukerTokenInfo)
-
-                        if (trygdetid != null) {
-                            trygdetidService.reberegnUtenFremtidigTrygdetid(
-                                behandlingId,
-                                trygdetid.id,
-                                brukerTokenInfo,
-                            )
-                            call.respond(
-                                trygdetidService.hentTrygdetidIBehandlingMedId(behandlingId, trygdetidId(), brukerTokenInfo)!!.toDto(),
-                            )
-                        } else {
-                            call.respond(HttpStatusCode.NoContent)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
